update api routes
