+++ src/api/spreadsheet.test.ts
@@ -1,8 +1,8 @@
 import { beforeAll, describe, expect, jest, test } from '@jest/globals';
 import type { GoogleSpreadsheetWorksheet } from 'google-spreadsheet';
 import type { SpyInstance } from 'jest-mock';
-import type { FailedEntry, LoggedIn } from '.';
-import { loggedin_sheet_name, log_sheet_name } from '../../consts';
+import type { FailedEntry, LoggedIn } from '../types';
+import { loggedin_sheet_name, log_sheet_name } from '../consts';
 import { addHours, addHoursSafe, configureDrive, getSpreadsheet, updateLoggedIn } from './spreadsheet';
 import PQueue from 'p-queue';
