+++ src/types.ts
@@ -1,0 +1,15 @@
+export type FailedEntry = {
+    name: string;
+    timeIn: number;
+    timeOut: number;
+    activity: string;
+}
+export type LoggedIn = {
+    [key:string]:number
+}
+
+export type Member = {
+    name: string;
+    firstname: string;
+    img: string;
+}
